% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbox.R
\name{gbox}
\alias{gbox}
\title{Create a quick boxplot in ggplot.}
\usage{
gbox(var, group = NULL, fill = NULL, orient = "vertical")
}
\arguments{
\item{group}{(optional) Categorical variable by which to group the box plots.}

\item{fill}{A color for the numeric box plot (default is 'NULL').}

\item{numeric}{Variable for which to create the box plot.}

\item{orientation}{Allows user to indicate vertical or horizontal plotting.}
}
\value{
This function returns a ggplot box plot object.
}
\description{
This will graph a given numeric vector in a ggplot-style box plot with the
y-axis labeled after the input 'vector'.  OR this will graph side-by-side
box plots of a numeric vector grouped by a given categorical vector.
The user may indicate 'vertical' (default) or 'horizontal' plotting.
The 'scales' package is required to implement viridis_c() over the number of
unique categories in 'group'.
}
\examples{
## Create a box plot of 'var' over four categories in 'group'.
var <- rnorm(100)
group <- factor(rep(c("Yes", "No", "Maybe", "Probably"), each = 25))
gbox(var, group, orient = 'horizontal', fill = 'orange')

}
